Update createDirAndFiles and deleteFils
This will check for / in the file name

+++ utils/deleteFiles.cjs
@@ -4,13 +4,17 @@ const path = require("path");
 function deleteFiles(dirName, files) {
   return new Promise((resolve, reject) => {
     let delteFilesPromises = files.map((file) => {
-      let filePath = path.join(
-        __dirname,
-        "userCreatedFiles/",
-        dirName,
-        file + ".json"
-      );
-      return fsPromises.unlink(filePath);
+      if (typeof file === "string" && file.includes("/")) {
+        return Promise.reject(`${file} not a string or contains /`);
+      } else {
+        let filePath = path.join(
+          __dirname,
+          "userCreatedFiles/",
+          dirName,
+          file + ".json"
+        );
+        return fsPromises.unlink(filePath);
+      }
     });
     Promise.allSettled(delteFilesPromises)
       .then((settledPromises) => {
